+++ src/impls.rs
@@ -1,5 +1,6 @@
 mod add_iterable;
 mod add_self;
+mod extend;
 mod index;
 mod intersection;
 mod sub_iterable;
